appone/requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
numpy==1.21.6
opencv-contrib-python==4.7.0.72
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
psycopg2-binary==2.8.5
